requirements.txt
pyfiglet
pytest
pycryptodomepytest
pycryptodome
hashlib
os